MaterialX: fix bug in thin-walled subsurface (#251)
As pointed out by Brian Sharpe of SideFX, there is a bug in the thin-walled subsurface, since `specular_color` is accounted for in both the BSDF weights and as an explicit multiplier.

Also, the graph can be simplified a bit using the mix operation with a more carefully chosen weight.